requirements.txt
fastapi==0.111.0
uvicorn==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
pyyaml==6.0.2
tensorflow-cpu==2.16.2
kaggle==1.6.17
